+++ tests/integration/test_bayesian_optimization.py
@@ -34,7 +34,7 @@ from trieste.acquisition.rule import (
 )
 from trieste.bayesian_optimizer import BayesianOptimizer
 from trieste.data import Dataset
-from trieste.models import GaussianProcessRegression
+from trieste.models.gpflow import GaussianProcessRegression
 from trieste.objectives import (
     BRANIN_MINIMIZERS,
     BRANIN_SEARCH_SPACE,
