+++ tests/integration/test_constrained_bayesian_optimization.py
@@ -23,7 +23,7 @@ from trieste.acquisition.function import ExpectedConstrainedImprovement, Probabi
 from trieste.acquisition.rule import EfficientGlobalOptimization
 from trieste.bayesian_optimizer import BayesianOptimizer
 from trieste.data import Dataset
-from trieste.models import GaussianProcessRegression
+from trieste.models.gpflow import GaussianProcessRegression
 from trieste.space import Box
 from trieste.utils import map_values
