+++ tests/unit/acquisition/test_function.py
@@ -25,6 +25,7 @@ import pytest
 import tensorflow as tf
 import tensorflow_probability as tfp
 
+from tests.util.acquisition.sampler import PseudoBatchReparametrizationSampler
 from tests.util.misc import (
     TF_DEBUGGING_ERROR_TYPES,
     ShapeLike,
@@ -35,8 +36,7 @@ from tests.util.misc import (
     random_seed,
     various_shapes,
 )
-from tests.util.model import GaussianProcess, QuadraticMeanAndRBFKernel, rbf
-from tests.util.sampler import PseudoBatchReparametrizationSampler
+from tests.util.models.gpflow.models import GaussianProcess, QuadraticMeanAndRBFKernel, rbf
 from trieste.acquisition.function import (
     GIBBON,
     AcquisitionFunction,
@@ -69,7 +69,7 @@ from trieste.acquisition.function import (
 )
 from trieste.data import Dataset
 from trieste.models import ProbabilisticModel
-from trieste.objectives.single_objectives import BRANIN_MINIMUM, branin
+from trieste.objectives import BRANIN_MINIMUM, branin
 from trieste.space import Box
 from trieste.type import TensorType
 from trieste.utils import DEFAULTS
