+++ tests/unit/acquisition/test_rule.py
@@ -22,7 +22,7 @@ import pytest
 import tensorflow as tf
 
 from tests.util.misc import empty_dataset, quadratic, random_seed
-from tests.util.model import QuadraticMeanAndRBFKernel
+from tests.util.models.gpflow.models import QuadraticMeanAndRBFKernel
 from trieste.acquisition import (
     AcquisitionFunction,
     AcquisitionFunctionBuilder,
