+++ tests/util/models/gpflow/test_models.py
@@ -1,13 +1,24 @@
-# Copyright (C) PROWLER.io 2021 - All Rights Reserved
-# Unauthorised copying of this file, via any medium is strictly prohibited
-# Proprietary and confidential
+# Copyright 2021 The Trieste Contributors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import numpy.testing as npt
 import pytest
 import tensorflow as tf
 import tensorflow_probability as tfp
 
 from tests.util.misc import quadratic, random_seed
-from tests.util.model import GaussianProcess
+from tests.util.models.gpflow.models import GaussianProcess
 
 
 def _example_gaussian_process() -> GaussianProcess:
