+++ trieste/models/gpflow/interface.py
@@ -22,8 +22,8 @@ from gpflow.models import GPModel
 
 from ...data import Dataset
 from ...type import TensorType
-from ..optimizer import Optimizer
 from ..interfaces import ProbabilisticModel
+from ..optimizer import Optimizer
 from .utils import module_deepcopy
